benchmarks/requirements.txt
noise
git+https://github.com/plottertools/pnoise#egg=pnoise
pytest-benchmark